requirements.txt
# tests
pytest==8.3.2
# formatter, linter, imports sorter
ruff==0.5.7
# pre-commit
pre-commit==3.8.0
# DEV deps
python-telegram-bot==21.7
rethinkdb
asyncio
mmh3==4.1.0
polars==0.19.19
simplejson==3.19.2
more-itertools==10.1.0
python-dotenv
sqlalchemy==2.0.36
openai==1.55.3
json_repair==0.30.2
requests==2.32.3
asyncpg==0.30.0
loguru==0.7.2
fastapi==0.115.5
pydantic-settings==2.6.1
gunicorn==23.0.0
uvicorn==0.32.1
envyaml==1.10.211231
langchain==0.3.9
langchain-openai==0.2.10
dotmap==1.3.30